create module fade.js
