+++ src/animation.js
@@ -1,5 +1,3 @@
-import {removeActiveClass} from "./helpers";
-
 /**
  * Scroll into view
  * @param target
@@ -12,85 +10,31 @@ export function scrollIntoView(target){
 
 
 /**
- * Fade
+ * Get margin height of element
+ * @param el
+ * @returns {number}
  */
-
-export function destroyFade(context){
-    context.wrapper.querySelectorAll(context.options.receiver).forEach(target => {
-        setCSS(target, {
-            opacity: '',
-            visibility: '',
-            position: '',
-            inset: ''
-        });
-        setCSS(target.parentElement, {
-            height: '',
-            position: ''
-        });
-
-        removeTransition(target);
-        removeTransition(target.parentElement);
-    });
-
-    removeActiveClass(context);
-}
-
-export function fadeOut(target, duration = 500, fn){
-    // before
-    setTransition(target, duration);
-
-    // animate
-    setCSS(target, {
-        opacity: '0',
-        visibility: 'hidden'
-    });
-
-    // end
-    setTimeout(() => {
-        removeTransition(target);
-
-        // callback
-        if(typeof fn === 'function') fn();
-    }, duration);
-}
-
-export function fadeIn(target, duration = 500, fn){
-    // before
-    setTransition(target, duration);
-    setTransition(target.parentElement, duration);
-
-    // animate
-    setCSS(target, {
-        opacity: '1',
-        visibility: 'visible'
-    });
-
-    // update parent height
-    setCSS(target.parentElement, {height: `${target.scrollHeight + getElementMarginHeight(target)}px`});
-
-    // end
-    setTimeout(() => {
-        removeTransition(target);
-        removeTransition(target.parentElement);
-
-        // callback
-        if(typeof fn === 'function') fn();
-    }, duration);
-}
-
-/**
- * Helpers
- */
-
 export function getElementMarginHeight(el){
     const computedStyle = getComputedStyle(el);
     return parseInt(computedStyle.getPropertyValue('margin-top') + computedStyle.getPropertyValue('margin-bottom'), 10);
 }
 
+
+/**
+ * Set CSS
+ * @param target
+ * @param props
+ */
 export function setCSS(target, props){
     Object.assign(target.style, props);
 }
 
+
+/**
+ * Set transition
+ * @param target
+ * @param duration
+ */
 export function setTransition(target, duration){
     setCSS(target, {
         transitionProperty: "height, margin, padding, opacity",
@@ -99,6 +43,11 @@ export function setTransition(target, duration){
     });
 }
 
+
+/**
+ * Remove transition
+ * @param target
+ */
 export function removeTransition(target){
     setCSS(target, {
         overflow: '',
